AnalyzeType ve AnalyzeDetail eklendi

+++ Analysis.Business/Abstract/IAnalyzeDetailManager.cs
@@ -4,7 +4,7 @@ namespace Analysis.Business.Abstract
 {
     public interface IAnalyzeDetailManager : IManager<AnalyzeDetail, int>
     {
-        public void SetLimit(AnalyzeDetail analyzeDetail, int limit, string analysisType);
+        public void SetLimit(AnalyzeDetail analyzeDetail, double limit, string analysisType);
 
     }
